Minor refactoring
Created wrapper component to reduce the div usage in the return statement.
Imported Fragment from the React component as a replacement to the div usage.
Imported Portal module from the React-dom to move the modal to a separate ID in the index.html.
Imported ref module from React to replace the useState functions used in the addUser.js module.